Added searchBar to ListVIew and changed ListViewController title
