+++ solutions/devsprint-bruno-ramos-5/GitHubApp/AppDelegate/SceneDelegate.swift
@@ -16,8 +16,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let windowScene = (scene as? UIWindowScene) else { return }
 
         self.window = UIWindow(frame: UIScreen.main.bounds)
-        let viewController = SampleConfigurator.make()
-        self.window?.rootViewController = UINavigationController(rootViewController: viewController)
+        self.window?.rootViewController = UINavigationController(rootViewController: ListViewController())
         self.window?.windowScene = windowScene
         self.window?.makeKeyAndVisible()
     }
