+++ solutions/devsprint-bruno-ramos-5/GitHubApp/Screens/List/ListViewController.swift
@@ -27,9 +27,8 @@ final class ListViewController: UIViewController {
     }
 
     override func viewDidLoad() {
-
         self.navigationController?.navigationBar.prefersLargeTitles = true
-        self.navigationItem.title = "GitHub App 🐙"
+        self.navigationItem.title = "Repositories"
     }
 
     override func viewDidAppear(_ animated: Bool) {
