+++ solutions/devsprint-bruno-ramos-5/GitHubApp/Screens/Protocols/ViewCode.swift
@@ -1,0 +1,23 @@
+//
+//  ViewCode.swift
+//  GitHubApp
+//
+//  Created by Vitor Conceicao on 06/07/22.
+//
+
+import Foundation
+
+protocol ViewCode {
+    func setupSubviews()
+    func setupConstraint()
+    func setupExtraConfiguration()
+    func setup()
+}
+
+extension ViewCode {
+    func setup() {
+        setupSubviews()
+        setupConstraint()
+        setupExtraConfiguration()
+    }
+}
